Refactor some more parts of the code, fix networking on Linux, fix sprite loading

+++ src/Common/server_addr.h
@@ -2,8 +2,8 @@
 #define DOLORI_COMMON_SERVER_ADDR_H_
 
 struct SERVER_ADDR {
-  unsigned long ip;
-  short port;
+  unsigned int ip;
+  unsigned short port;
   unsigned char name[0x14];
   unsigned short usercount;
   unsigned short state;
